Use Browserify in build (without require()s yet)
I am committing this stage of the adding Browserify as a commit so that I can return to it later when I want to debug source maps.

Source maps currently aren’t working completely, and I haven’t found an easy way to fix that – it might be problems that need to be fixed in dependencies or in the browser. Debugging source maps will get even harder after I add require()s, so this commit is the easiest state from which I can debug source maps.