Move deploy script from package.json to Makefile
I had forgotten that the Makefile existed. It seems like a better place to put scripts – I can write the scripts on multiple lines there.

I renamed `server` to `serve` in the Makefile. I’m used to Broccoli’s wording now, and using the verb makes it clearer that I’m not building a server when I run `make`.

In the README, I now refer only to Makefile scripts, when a script exists.