Normalize relative weights after slider drags
I don’t need to normalize the weights after user input in the text fields – just the sliders.

This does make it harder for the user to give all attributes whole-number weights, such as [3, 2]. They can do it, but it will be undone if they use the sliders. But I don’t see a way around that without also allowing the numbers to grow or shrink way too much when repeatedly dragging alternate sliders right or left.

There is one thing that could still be improved about this. I want to round the relative weights to two decimal places when they are displayed in the text field, so the users won’t see the repeating decimals resulting from the division.