Make the matrix react to choice/attribute changes
Rating input and score calculation are not implemented yet.

I had to extract the storing of choices and attributes into a parent controller, DMCController. The sub-controllers currently look in the $scope to get the collections they used to modify. I will eventually look for a better way of doing this that makes the dependency more explicit.